requirements.txt
bcrypt==4.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
getenv==0.2.0
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlalchemy==2.0.36; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'